chore: add managePrimaryEmailForUser file
