+++ src/app/api/member/actions/index.ts
@@ -349,82 +349,6 @@ async function updateMemberMissions(
     revalidatePath("/community/[id]", "layout");
 }
 
-export async function managePrimaryEmailForUser({
-    username,
-    primaryEmail,
-}: {
-    username: string;
-    primaryEmail: string;
-}): Promise<void> {
-    const session = await getServerSession(authOptions);
-    if (!session || !session.user.id) {
-        throw new AuthorizationError();
-    }
-    const isCurrentUser = session?.user.id === username;
-    const user = await userInfos({ username }, isCurrentUser);
-    if (!user.authorizations.canChangeEmails) {
-        throw new AuthorizationError();
-    }
-    const primaryEmailIsPublicServiceEmail = await isPublicServiceEmail(
-        primaryEmail
-    );
-    if (!primaryEmailIsPublicServiceEmail) {
-        throw new BusinessError(
-            `L'email renseigné n'est pas un email de service public`
-        );
-    }
-    if (isAdminEmail(primaryEmail)) {
-        throw new AdminEmailNotAllowedError();
-    }
-
-    if (user.userInfos.primary_email?.includes(config.domain)) {
-        await betagouv.createRedirection(
-            user.userInfos.primary_email,
-            primaryEmail,
-            false
-        );
-        try {
-            await betagouv.deleteEmail(
-                user.userInfos.primary_email.split("@")[0]
-            );
-        } catch (e) {
-            console.log(e, "Email is possibly already deleted");
-        }
-    } else {
-        try {
-            await mattermost.getUserByEmail(primaryEmail);
-        } catch {
-            throw new BusinessError(
-                `L'email n'existe pas dans mattermost, pour utiliser cette adresse comme adresse principale ton compte mattermost doit aussi utiliser cette adresse.`
-            );
-        }
-    }
-    await db
-        .updateTable("users")
-        .where("username", "=", username)
-        .set({
-            primary_email: primaryEmail,
-            username,
-        })
-        .execute();
-
-    await addEvent({
-        action_code: EventCode.MEMBER_PRIMARY_EMAIL_UPDATED,
-        created_by_username: session.user.id,
-        action_on_username: username,
-        action_metadata: {
-            value: primaryEmail,
-            old_value: user
-                ? user.userInfos.primary_email || undefined
-                : undefined,
-        },
-    });
-
-    console.log(`${session?.user.id} a mis à jour son adresse mail primaire.`);
-    revalidatePath("/community/[id]", "layout");
-    return;
-}
-
 export async function deleteEmailForUser({ username }: { username: string }) {
     const session = await getServerSession(authOptions);
     if (!session || !session.user.id) {
@@ -570,10 +494,7 @@ export const safeUpdateCommunicationEmail = withErrorHandling<
     UnwrapPromise<ReturnType<typeof updateCommunicationEmail>>,
     Parameters<typeof updateCommunicationEmail>
 >(updateCommunicationEmail);
-export const safeManagePrimaryEmailForUser = withErrorHandling<
-    UnwrapPromise<ReturnType<typeof managePrimaryEmailForUser>>,
-    Parameters<typeof managePrimaryEmailForUser>
->(managePrimaryEmailForUser);
+
 export const safeDeleteEmailForUser = withErrorHandling<
     UnwrapPromise<ReturnType<typeof deleteEmailForUser>>,
     Parameters<typeof deleteEmailForUser>
