+++ src/components/MemberPage/Email/BlocConfigurerEmailPrincipal.tsx
@@ -4,14 +4,9 @@ import { fr } from "@codegouvfr/react-dsfr";
 import Accordion from "@codegouvfr/react-dsfr/Accordion";
 import Button from "@codegouvfr/react-dsfr/Button";
 import Input from "@codegouvfr/react-dsfr/Input";
-import axios from "axios";
 
-import {
-    managePrimaryEmailForUser,
-    safeManagePrimaryEmailForUser,
-} from "@/app/api/member/actions";
+import { safeManagePrimaryEmailForUser } from "@/app/api/member/actions/managePrimaryEmailForUser";
 import { memberBaseInfoSchemaType } from "@/models/member";
-import routes, { computeRoute } from "@/routes/routes";
 
 export default function BlocConfigurerEmailPrincipal({
     canChangeEmails,
